+++ src/common/constants.ts
@@ -1,0 +1,4 @@
+const DOMAIN = "@sbx";
+const OTP_LENGTH = 6;
+
+export { DOMAIN, OTP_LENGTH };
