+++ src/components/auth/AbhaNumberOtpFlow.tsx
@@ -26,9 +26,9 @@ import OtpInput from "@/components/auth/ui/otp-resend-input";
 
 import { useOtpFlow } from "@/hooks/useOtpFlow";
 
-import { FormMemory } from "@/types/auth";
+import { OTP_LENGTH } from "@/common/constants";
 
-const OTP_LENGTH = 6;
+import { FormMemory } from "@/types/auth";
 
 type AbhaNumberOtpFlowProps = {
   flowType: "login" | "enrollment";
